// typechain-types/BaseV1Router01.ts
export declare namespace BaseV1Router01 {
  export type RouteStruct = { from: string; to: string; stable: boolean };

  export type RouteStructOutput = [string, string, boolean] & {
    from: string;
    to: string;
    stable: boolean;
  };
}