// typechain-types/BaseV1GaugeFactory.ts
export interface BaseV1GaugeFactoryInterface extends utils.Interface {
  contractName: "BaseV1GaugeFactory";
  functions: {
    "createGauge(address,address,address)": FunctionFragment;
    "createGaugeSingle(address,address,address,address)": FunctionFragment;
    "last_gauge()": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "createGauge",
    values: [string, string, string]
  ): string;
  encodeFunctionData(
    functionFragment: "createGaugeSingle",
    values: [string, string, string, string]
  ): string;
  encodeFunctionData(
    functionFragment: "last_gauge",
    values?: undefined
  ): string;

  decodeFunctionResult(
    functionFragment: "createGauge",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "createGaugeSingle",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "last_gauge", data: BytesLike): Result;

  events: {};
}

export interface BaseV1GaugeFactory extends BaseContract {
  contractName: "BaseV1GaugeFactory";
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: BaseV1GaugeFactoryInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    createGauge(
      _pool: string,
      _bribe: string,
      _ve: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<ContractTransaction>;

    createGaugeSingle(
      _pool: string,
      _bribe: string,
      _ve: string,
      _voter: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<ContractTransaction>;

    last_gauge(overrides?: CallOverrides): Promise<[string]>;
  };

  createGauge(
    _pool: string,
    _bribe: string,
    _ve: string,
    overrides?: Overrides & { from?: string | Promise<string> }
  ): Promise<ContractTransaction>;

  createGaugeSingle(
    _pool: string,
    _bribe: string,
    _ve: string,
    _voter: string,
    overrides?: Overrides & { from?: string | Promise<string> }
  ): Promise<ContractTransaction>;

  last_gauge(overrides?: CallOverrides): Promise<string>;

  callStatic: {
    createGauge(
      _pool: string,
      _bribe: string,
      _ve: string,
      overrides?: CallOverrides
    ): Promise<string>;

    createGaugeSingle(
      _pool: string,
      _bribe: string,
      _ve: string,
      _voter: string,
      overrides?: CallOverrides
    ): Promise<string>;

    last_gauge(overrides?: CallOverrides): Promise<string>;
  };

  filters: {};

  estimateGas: {
    createGauge(
      _pool: string,
      _bribe: string,
      _ve: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<BigNumber>;

    createGaugeSingle(
      _pool: string,
      _bribe: string,
      _ve: string,
      _voter: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<BigNumber>;

    last_gauge(overrides?: CallOverrides): Promise<BigNumber>;
  };

  populateTransaction: {
    createGauge(
      _pool: string,
      _bribe: string,
      _ve: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<PopulatedTransaction>;

    createGaugeSingle(
      _pool: string,
      _bribe: string,
      _ve: string,
      _voter: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<PopulatedTransaction>;

    last_gauge(overrides?: CallOverrides): Promise<PopulatedTransaction>;
  };
}